% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mass_waste.R
\name{make_mass_waste}
\alias{make_mass_waste}
\title{make_mass_waste
Get the overall average by taking the average of the stand averages. This
means that each stand gets the same weight (/#of stands) in the calculation.
When multiplier is specified, the average is calculated per "multiplier",
for example per capita.}
\usage{
make_mass_waste(
  df,
  group = "",
  idvar = "ID",
  dataset = "HWS2022",
  indicatorOut = TRUE,
  measure = c("MASS", "VOL")[1],
  multiplier = c(NA, "number_members_residing")[1],
  pername = c("household", "capita")[1],
  unt = c("kg", "l")[1],
  period = "WEEK",
  srcpth = "",
  ...
)
}
\arguments{
\item{df}{tibble}

\item{group}{Charcter.}

\item{idvar}{Charcter.}

\item{dataset}{Charcter.}

\item{indicatorOut}{Logic}

\item{measure}{Charcter.}

\item{multiplier}{Charcter. Defalut NA}

\item{pername}{Charcter.}

\item{unt}{Charcter.}

\item{period}{Charcter.}

\item{srcpth}{Charcter.}

\item{category}{Currently ignored - for campatability with other functions in family to enable uniform arg list}

\item{namewash}{Currently ignored - for campatability with other functions in family to enable uniform arg list}
}
\description{
make_mass_waste
Get the overall average by taking the average of the stand averages. This
means that each stand gets the same weight (/#of stands) in the calculation.
When multiplier is specified, the average is calculated per "multiplier",
for example per capita.
}
\examples{
make_mass_waste(df = ls_ana$data[[1]], multiplierdf = ls_bd$data[[1]], pername = "capita", group = "CAP")


}
