% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_measure_waste.R
\name{make_measure_waste}
\alias{make_measure_waste}
\title{make_measure_waste_cat}
\usage{
make_measure_waste(
  loopdf,
  loopname = NULL,
  bycat = c(NA_character_, "BYCAT_", "BYINTVCAT")[1],
  measure = c("MASS", "VOLUME")[1],
  group = c("HH"),
  category = c("", "RCLB", "NONRCLB")[1],
  varname = "variable",
  idvar = c("instanceid", "ID")[1],
  valuename = "value",
  dataset = "",
  namewash = c("", "waste_type_sample_weight_")[1],
  pername = c("", "household", "capita")[1],
  period = "WEEK",
  unt = "kg",
  srcpth = "",
  df = NULL,
  multiplier = NULL,
  digits = 4,
  ...
)
}
\arguments{
\item{loopdf}{List of data.frames with results of measurement loop per category with a name and value}

\item{loopname}{Name in loopdf of the list item with the tibbe to use}

\item{bycat}{Character.}

\item{measure}{Character. Abbreviation for type of measure. "MASS" or "VOL"}

\item{group}{Character. Abbreviation  of level at which data is collected. "HH" for household}

\item{category}{Character.}

\item{varname}{Character. Name in loopdf with class}

\item{idvar}{Character.}

\item{valuename}{Character. Name in loopdf with value}

\item{dataset}{Character. Description of dataset}

\item{namewash}{Character. Portion of varname to remove, "" is nothing}

\item{pername}{Character. Description of level at which data is collected. Full text of group}

\item{period}{Character. Period over which measurement is averaged}

\item{unt}{Character. Unit of measure}

\item{srcpth}{Character. Source path}

\item{df}{NULL. For campatibility}

\item{multiplier}{NULL. For campatibility}
}
\description{
Make the mass per category indicator from the results of a looping question with measurements of each relevant category
}
