% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_to_intv2.R
\name{recode_to_intv2}
\alias{recode_to_intv2}
\title{recode_to_intv2}
\usage{
recode_to_intv2(
  df,
  idvar = "ID",
  instancevar = "instanceid",
  typepatt = c("waste_type_sample_weight_", "")[2],
  wideout = TRUE
)
}
\arguments{
\item{idvar}{character with variable containg id. Default: "ID"}

\item{instancevar}{character}

\item{typepatt}{character}

\item{wideout}{logical}

\item{l}{list}
}
\value{
tibble containing total waste per main categories
 (recyclable, non-recyclable, sanitary, organic)
}
\description{
recode_to_intv2
}
