% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_loop_cat.R
\name{summarise_loop_cat}
\alias{summarise_loop_cat}
\title{summarise_loop_cat}
\usage{
summarise_loop_cat(
  loop = df_ana$loops[[1]][[4]],
  patt = "",
  idvar = "ID",
  recordid = "instanceid",
  useMean = TRUE,
  wideout = TRUE
)
}
\arguments{
\item{loop}{tibble}

\item{patt}{regular expression}
}
\value{
tibble
}
\description{
Summarise a loop section by household and category including 0 if not present
}
\examples{
summarise_loop_cat(loop = df_ana$loops[[1]][[1]], patt = "waste_type_non_recycling_weight_")
}
