% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cat_JP.R
\name{summarise_loop_cat_JP}
\alias{summarise_loop_cat_JP}
\title{summarise_loop_cat}
\usage{
summarise_loop_cat_JP(
  df = df_main,
  patt = "",
  idvar = c("ID"),
  recordid = "instanceid",
  analysis_var = "tot_weight",
  summary_by = "place",
  useMean = TRUE,
  longformat = TRUE,
  summary_table = TRUE
)
}
\arguments{
\item{patt}{regular expression}

\item{loop}{tibble}
}
\value{
tibble
}
\description{
Summarise a loop section by household and category including 0 if not present
}
\examples{
summarise_loop_cat(loop = df_ana$loops[[1]][[1]], patt = "waste_type_non_recycling_weight_")
}
