% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{summarise_loop_cat_JP_test}
\alias{summarise_loop_cat_JP_test}
\title{summarise_loop_cat_JP_test}
\usage{
summarise_loop_cat_JP_test(
  df = df_main,
  patt = "",
  idvar = c("ID"),
  recordid = "instanceid",
  analysis_var = "tot_weight",
  summary_by = "place",
  useMean = TRUE,
  longformat = TRUE,
  summary_table = TRUE
)
}
\arguments{
\item{summary_table}{}
}
\value{
tibble
}
\description{
summarise_loop_cat_JP_test
}
