% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsummary.R
\name{tenpointsummary}
\alias{tenpointsummary}
\title{tenpointsummary}
\usage{
tenpointsummary(x)
}
\arguments{
\item{x}{}
}
\value{
Tibble
tenpointsummary

tibble
}
\description{
convenience function to summarise data inside dplyr pipeline
}
