% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_to_volume.R
\name{words_to_volume}
\alias{words_to_volume}
\title{words_to_volume}
\usage{
words_to_volume(
  x,
  words = c("Full", "3/4", "Half", "Quarter", "Less than a quarter"),
  weights = c(1, 0.75, 0.5, 0.25, 0.125),
  base = c(70, 20)[1]
)
}
\arguments{
\item{base}{}
}
\description{
words_to_volume
}
